+++ models/user.model.js
@@ -1,0 +1,15 @@
+const mongoose = require('mongoose')
+const Schema   = mongoose.Schema
+
+let userSchema = mongoose.Schema({
+  username: String,
+  firstname: String,
+  lastname: String,
+  password: String,
+}, {
+  timestamps: true
+})
+
+let user = mongoose.model('User', userSchema)
+
+module.exports = user
