+++ models/video.model.js
@@ -1,11 +1,13 @@
 const mongoose = require('mongoose')
 
 const videoSchema = mongoose.Schema({
-    title: String,
-    url: String,
-    description: String,
-    like: Number,
-    dislike: Number
+  title: String,
+  url: String,
+  description: String,
+  like: Number,
+  dislike: Number
+}, {
+  timestamps: true
 })
 
 const Video = mongoose.model('Video', videoSchema)
