+++ package.json
@@ -1,23 +1,23 @@
 {
   "name": "server",
-  "version": "0.0.0",
-  "private": true,
+  "version": "1.0.0",
+  "description": "",
+  "main": "app.js",
   "scripts": {
-    "start": "nodemon ./bin/www"
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "start": "nodemon app.js"
   },
+  "author": "",
+  "license": "ISC",
   "dependencies": {
     "@google-cloud/storage": "^1.6.0",
     "bcryptjs": "^2.4.3",
-    "cookie-parser": "~1.4.3",
-    "debug": "~2.6.9",
+    "cors": "^2.8.4",
     "dotenv": "^5.0.1",
-    "express": "~4.16.0",
+    "express": "^4.16.3",
     "fb": "^2.0.0",
-    "http-errors": "~1.6.2",
-    "jade": "~1.11.0",
     "jsonwebtoken": "^8.2.1",
-    "mongoose": "^5.0.17",
-    "morgan": "~1.9.0",
+    "mongoose": "^5.0.18",
     "multer": "^1.3.0"
   }
 }
