+++ routes/index.js
@@ -1,7 +1,7 @@
-var express = require('express');
-var router = express.Router();
+var express = require('express')
+var router = express.Router()
 
 /* GET home page. */
-router.get('/');
+router.get('/')
 
-module.exports = router;
+module.exports = router
