+++ routes/users.js
@@ -1,9 +1,9 @@
-const express = require('express');
-const router = express.Router();
-const { loginUser, registerUser } = require('../controllers/user_controller.js')
+const express = require('express')
+const router = express.Router()
+const { loginUser, registerUser } = require('../controllers/user.controller')
 
-router.post('/login', loginUser);
+router.post('/login', loginUser)
 
 router.post('/register', registerUser)
 
-module.exports = router;
+module.exports = router
